fix: remove unnecessary type casts and add Cyrillic layout support for terminal
- Remove unnecessary u64 casts in statvfs operations (clippy fix)
- Add translate_hotkey() support to terminal for Cyrillic layout
- Now Ctrl+Shift+М works same as Ctrl+Shift+V for paste in terminal

Fixes GitHub Actions clippy errors blocking release build.

+++ src/panels/file_manager/file_info.rs
@@ -346,8 +346,8 @@ impl FileManager {
                 // f_bavail - available blocks for non-privileged users
                 // f_blocks - total blocks in the filesystem
                 // f_bsize - block size in bytes
-                let available = (stat.f_bavail as u64) * (stat.f_bsize as u64);
-                let total = (stat.f_blocks as u64) * (stat.f_bsize as u64);
+                let available = stat.f_bavail * stat.f_bsize;
+                let total = stat.f_blocks * stat.f_bsize;
 
                 Some(DiskSpaceInfo {
                     device,
